Fixing some errors in the tests (caused by a previous commit).

+++ blueprints-test/src/test/java/com/tinkerpop/blueprints/util/wrappers/event/EventGraphTest.java
@@ -434,7 +434,7 @@ public class EventGraphTest extends GraphTest {
         Vertex vertex = createVertex();
         vertex.setProperty("test", 123);
 
-        assertEquals(12345, (long) vertex.getProperty("setInListener"));
+        assertEquals(12345, (int) vertex.getProperty("setInListener"));
         assertEquals(2, listener.vertexPropertyChangedEventRecorded());
     }
 }
